decouple MethodCallToVariableNameResolver
